Implemented settings menu
The settings menu successfully takes in user supplied IP address and
port, and is used in sending the UDP packet. Pressing 'done' on the
keyboard will unfocus the TextView and close the keyboard.
Something to consider changing is every time the app sends a command it
must parse the settings. Also, it would be nice to be able to save the
settings for a quick load.