Added HMAC, ACK, and retries
The android app will resend commands until it either receives an ACK or
it surpasses MAX_RETRIES (3). The server has a history of timestamps, so
if it receives the same command it will simply resend the ACK. It only
executes the command once. Both the server and client must use the same
password.
New features:

    * HMAC
    * Retries / timeouts
    * History
    * Password (for HMAC)
    * Cursor position at the end of EditText