FF / REW states, callbacks, and image buttons
The UI now uses ImageButtons instead of text and callbacks once the
UDPPacket completes. I created a new class for this that handles it
automatically. It works well for play / pause.  There is some
redundancy in the code with the Callback interface in MainActivity.
Also, fast-forward and rewind states were added so the user can hold
down the button to FF / REW. Unfortunately, the way the server is setup
and how I implemented the states, it sends an extra ACK. To fix this I
will properly implement states.