Implemented preliminary version of LibraryActivity
LibraryActivity uses a LoaderManager to fetch the data from the server
in the background while a "Loading..." dialog displays. In this commit
there is no communication with the server and it simply sleeps for 2
seconds to show it is working.
This commit also includes a DirectoryListing class which is contains an
ArrayList of files and directories and a fetch() method for filling
these arrays.