Automatic speedup for repeat commands
Unless a delay is specified, the repeat command will automatically
speedup (decrease the period) in which it calls the command to be
repeated. To avoid many, many calls in the case of "seek", the function
will modify the arguments to seek faster (double every 3 seconds).
Otherwise, it halves the delay every 3 seconds.