Added auto search for server
This required modifying the server to respond to a health command
without validating, because the app does not have the password when it
starts.

An alternate approach would be to search for the server once the
password is entered, but I think relaxing the security for health
commands is fine.