Add server run framework
If daemon is set, the run function will fork a new process to receive
commands and return control back to the caller. It then stops the server
by sending it a kill signal.

This commit also introduces a _serve function to handle the already
authenticated command. At this point it simply sends an ACK with "Not
Implemented" message.