Add basic _serve function
The _serve function is a switch case looking for specific commands. This
approach is more rigid than the previous implementation which allowed
the client to specify any properties to get or set and whitelisted some
other commands (e.g. loadfile).

The design boils down to tradeoffs between flexibility, security, and
abstraction. Because the backend controller is abstracted by the
MediaController class and its functions, it is best to take the approach
in this commit as not all backends will have a set / get property.