Lower history_size to 4
Due to the nature of the server, the history is only used when the
client does not receive a response from the server in time. This rarely
happens, so a small history size is acceptable for caching that
situation.

As an aside, the history feature is not really necessary if all
controller actions are idempotent. At this point, the only
state dependent action is seek.